[OPIK-8284] Review: one namespace per entity, and a registry that is only a table
@alexkuzmik asked for `read_list` to be modular the way `writes/` was made
modular in #186: every entity's logic in its own sub-namespace, the root
holding dispatch, registration and what every entity shares, and adding an
entity being a namespace plus an import. He was right, and the registry was
the evidence: 995 lines, of which about 900 were thirteen entities' fetchers,
list functions, name searches and compressors.

`registry.py` is 93 lines now. It imports thirteen handlers, keys them by
entity type, and holds the two behaviours that read the table rather than a
row: the alias an `entity_type` resolves through, and the compressor an entity
did not override.

Each entity declares its own handler beside its own code, in `entities/`. A
module where one file is enough (span, experiment, prompt with its versions,
test_suite with its items, score_name, online_rule, trace, thread) and a
package where it is not:

- `project/` has the read that fans out plus `summary`, `vocabulary` and
  `contents`, one file per part of the answer.
- `project_metric/` has `catalog` (what can be asked, and every refusal that
  needs no call), `runner` (the order of operations), `table` (how an answer
  reads) and `reference` (what `schema()` says).
- `agent_insights_issue/` is the Diagnostics package, moved under the entity
  it belongs to, with the entity's own read and list joining `state` and
  `availability`.

Three things came out of the root along the way, because more than one entity
needs them: `handler.py` for the contract (`diagnostics/state.py` was reaching
into the registry for `PageContext` under TYPE_CHECKING, which is the
circularity this removes), `paging.py` for reading the backend's page envelope,
and `unsupported.py` for the two sentinels that belong to no entity.

The dispatchers no longer name an entity. Two branches on `entity_type` are
now hooks on the handler, the way the write registry carries `validate_fn` and
its siblings: `run_fn` for an entity that answers `list` whole (a time series
is not a collection), and `reference_fn` for one whose schema is a catalog
instead of a field table. The one remaining literal was a nicer sentence for
why an experiment takes no window; that reason now sits in `oql.py` beside the
list of which entities do take one.

`test_modular.py` pins the shape: the root's module list, that every
registered entity has a namespace, that the registry defines only its two
table-level functions and nothing async, that neither dispatcher branches on
an entity name, and that the readable and listable surfaces are unchanged by
the move. The property is cheap to state and expensive to rediscover.

1756 tests, ruff and mypy clean. Behaviour is identical: no entry lost a
column, a flag or a description, because the handler literals were lifted from
the registry's own source rather than retyped. Re-verified live against
www.comet.com through the packaged wheel.
