+++ src/opik_mcp/read_list/reference.py
@@ -24,7 +24,7 @@ from opik_mcp.read_list.oql import (
     SUPPORTED_ENTITIES,
     WINDOWED_ENTITIES,
 )
-from opik_mcp.read_list.project_metrics import reference as metric_reference
+from opik_mcp.read_list.registry import ENTITY_REGISTRY
 from opik_mcp.read_list.sorting import SORT_FORM, sortable_names
 
 LIST_SCHEMA_KEYS: Final[tuple[str, ...]] = (
@@ -57,13 +57,12 @@ FILTER_EXAMPLES: Final[dict[str, tuple[str, str]]] = {
 
 def list_reference(entity_type: str) -> dict[str, Any]:
     """The ``schema("list.<entity>")`` payload. ``entity_type`` must be supported."""
-    if entity_type == "project_metric":
-        # Its own tables (metrics, intervals, limits) rather than OQL fields;
-        # the filter fields are those of whichever entity the metric is about.
-        # The reference lives beside the catalog it describes — it is that
-        # data's own documentation, and moving it here would only move the
-        # import with it.
-        return metric_reference()
+    handler = ENTITY_REGISTRY.get(entity_type)
+    if handler is not None and handler.reference_fn is not None:
+        # The entity documents itself: a metric's reference is its catalog of
+        # metrics, intervals and limits, not a table of OQL fields. It lives
+        # beside the data it describes.
+        return handler.reference_fn()
     fields: dict[str, dict[str, Any]] = {}
     for name, ftype in FILTERABLE_FIELDS[entity_type].items():
         spec: dict[str, Any] = {"type": ftype, "operators": list(OPERATORS_BY_TYPE[ftype])}
